Fix unrenamed API parameter for intervention field.
The old prevention API parameter has been dropped both in the SP as well
as the call made in Zebra, the API did not have this change, causing it
to continue to look for the `prevension` parameter (which will always be
an empty string). As a result, the actual intervention search parameter
is never passed to the stored procedure.

Fixes PT-246.
Merged in bugfix/PT-246-api-prevension-mismatch (pull request #20)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>