Implement new User Creation page for Admins.
Created a new set of pages for the new user registration workflow, which
is similar to the original workflow, but also requires user to set a
password when they confirm their email.

The new page uses the new WebApi UserController to create the user. Old
implementations (in many locations) have been marked as deprecated to
centralize user creation to a single API location.
AddED callback URL changes and new page when following link.

Added a new email (RegistrationEmail) that is sent for this new
workflow.

Fixed invalid UserManager in the Api Controllers; this is now
initialized properly from OwinContext, so an additional NuGet package
was installed.
Fixed lint warnings and refactored code into the Library project.

Part of PT-234.
Merged in feature/PT-234-confirm-email-changes (pull request #23)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>