Implement updates to Reset Password UI.
Added new API in the UserController to support reset password workflows.
The code/token from the URL is used for verification for the calls made
during this workflow.

Updated the existing Reset password to also use the new API.
Moved the ResetPasswordViewModel into a separate file.

Changed sending Welcome Email logic to be triggered from the client side
as an async JS call to not have to wait for the email to be sent before
showing the password setting page.

Also updated identity token expiry to be 7 days, with is configurable in
the config file.
Also fixed email footer to remove the word 'infectious'.
Also included Validation of AntiForgery token for WebAPI.

Part of PT-235.
Merged in feature/PT-235-reset-password-ui-update (pull request #33)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>