Implement the APIs to retrieve data for group by disease.
Updated the GetEventCasePartial API to return different partial views
based on the grouping type. If the grouping type is Group by Disease,
then return the newly created Partial View.

Re-implemented the original behaviours for the old Event Case List:
- clicking item opens the event details
- setting global variables for filter results

Updated the Group by Disease drop-down list option to be disabled if no
AOIs were used in the filter.

Added new API for the client to make asynchronous calls to retrieve the
disease Case Count and Travellers data. This API calls the 2 SPs and
return them in a single object for a given disease and set of geoname ids.

Moved Risk levels logic to a helper function. This determines based on
the Max probability which category of low, medium, high it falls into.

Updated EF to include the 2 SPs used.

Part of PT-179 and PT-190.
Merged in feature/PT-190-disease-groupby-data-api (pull request #53)
Approved-by: Gursharn Singh <gru@bluedot.global>