Fix issue with suggested events not being deleted
A suggested event was not being deleted in the db because the parameter of the controller does not match the tag in the data passed by `_SuggestedEventView.cshtml`. The order of operations has also been changed so that only when the suggested event been successfully removed from the db will it be removed from the left panel list.

Resolves PT-366.

Approved-by: Basam Ahmad
Approved-by: Gursharn Singh
Approved-by: Sonya Karamchandani