Merged in feature/PT-374-execute-usp_insertactivegeonamesb (pull request #61)
Executing usp_InsertActiveGeonamesByGeonameIds to insert only the event locations that have been used by the events and articles only.

* Executing usp_InsertActiveGeonamesByGeonameIds to insert only the event locations that have been used by the events and articles only.
This SP is inserting the geonames into place.ActiveGeonames table in BiodZebra (Insights) database that been used by the articles and the events.
Expected results is the geonames records will be reduced from 4.8 million to around 5000 thousands geonames only. This means the calculation for the exportation and importation and all the calculations and getting the data will be faster because joining with 5000 thousands geonames is faster than joining with 4.8 million geonames records.

Approved-by: Gursharn Singh
PT-374