Implement logic for sub-grouping events in disease groups.
Previously, the group by disease only is done on diseases that are
returned by the existing event list (aka, no new events are added).
After clarification, events that have no risk individually, but after
aggregated with all events of the same disease that is of risk, needs to
be added and shown. This is implemented on the API side that modifies
the SP call parameters based on that logic.

Also implemented new logic to dynamically show or hide a disease group
on the client side, if the asynchronous request comes back that the
group is not supposed to be visible (no cases or too low risk).

Added no results template to the event list panel as well logic to
dynamically compute the number of events visible so that deferred logic
that ends up removing all the groups will trigger the no results to
show. Also updated the count to reflect the number of events that are
shown at the end, after the deferred calculations return.

Added helper function to create text for Traveller Interval, and
included unit tests for this function.

Part of PT-182.
Merged in feature/PT-182-subgrouping-disease-group (pull request #68)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>