Update logic to aggregate travellers before returning.
Previously, the aggregate count of travellers was retrieved with
aggregated case count from an API asynchronously from the client. Due to
faulty logic in aggregating the risk probabilities (average is
incorrect), the new logic will be that the grouping will only show if
there is at least 1 event that exceeds the risk threshold. Total
travellers will be aggregated by summing, and returned with the Event
List result, hence no longer needing the async call.

As a result, the API has been updated to only retrieve the total case
count. UI has been updated to not show a loading for the travellers, as
the value is already available.

Unit Tests were updated/fixed accordingly.
Also fixed the link to open Filter Panel in the empty event list not
implemented.

Part of PT-190.
Merged in feature/PT-190-aggreate-travellers (pull request #77)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>