Updated logic for filtering disease group.
Previously, the logic for hiding/showing depended on whether at least 1
event exceeds the threshold. This has been changed to a new calculation
method.

Also restored loading overlay to still apply to travellers.
Since the heading is not shown, the travellers shouldn't be shown, as
the section can still be hidden. Restored the loading overlay for the
travellers text, and made it only visible after return of the async
request.

Part of PT-104. and fixes PT-190.
Merged in feature/PT-104-aggr-risk-calculation (pull request #79)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>