Make saving email async call to db use different dbcontext.
Previously, a single DbContext was shared between the multiple threads
sending the emails, which caused an exception, as DbContext is not
thread-safe. A factory to return new DbContext instance was created and
injected into the Email flow, so that it can be replaced during testing.

Fixes PT-378.
Merged in bugfix/PT-378-email-save-async-bug (pull request #96)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>