Implement default group by to be by Disease.
Refactored the Group By logic into a separate helper class. Updated the
operations to be similar to Sort By in terms of determining which type
of grouping to use.

Added unit testing on the Group By helper function.

Updated accessors to use the helper function (which will query from the
database before falling back to the default). Also made the constant
private.

Part of PT-455.
Merged in feature/PT-455-default-group-by-disease (pull request #121)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>