Load setting from request cookie for sort/group options.
Updated the ZebraInit controller to accept sort/group option with
defaults so that if there is a setting to use, then use it otherwise
fall back to the default sort/group types.

Fixed the cookie code to store ID not the widget value, which was
currently storing the value of the drop down list pre-update.

Part of PT-455.
Merged in feature/PT-455-load-setting-from-cookie (pull request #126)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>