Update buttons in Filter Summary.
Updated the See all/custom events and modify filters button to the new
style in the Filter Summary. Also added a new collapse button that will
be used later. Aligned the blue tab icon and the zoom buttons to the
same height as the buttons in the Filter Summary.

Also updated Event details button style to match new buttons.

Part of PT-24.
Merged in feature/PT-24-filter-summary-buttons (pull request #141)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>