Implement nesting in csharp
EventCaseCountModel has been implemented as a tree structure to determine direct children. When the raw data is not available for all levels, this tree will be useful for determining which case counts will be nested to which level. For example, When Toronto, Ontario, Montreal, and Canada case counts are available, this will help in determining  that Montreal case counts should be nested to Canada's (if applicable).

EF has also been updated to add ActiveGeoname and Xtbl_Event_Location

Approved-by: Kevin Chu