Update airports to be independent from user AOI.
Destination Airports should be independent of the user AOI when shown on
the map (as if it is dependent, then the only airports that will be
shown are the ones near the user AOI, same for all events).

Made the Get Destination API call default to global view.
The geonames parameter is now made optional, with default as '-1' for
global view (all airports).

Part of PT-92.
Merged in feature/PT-92-airports-independent-of-aoi (pull request #148)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>