Add logic to store the Outbreak Potential results from George.
George API call can sometimes be slow if the service is not actively
serving requests (can be up to ~7s). Any calls to George to retrieve
Outbreak Potential Category will now be cached in the new table, and
retrieved from the table if found.

Fix ednullable importation risk in sorting.

Part of PT-743.
Merged in feature/PT-743-george-cache (pull request #246)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>