Extract out Sematic-UI from react app build.
Sematic-UI (not Semantic UI React) is being used to display popups (tooltips) on hover inside map. This is only use of Sematic-UI so far.

Sematic-UI build process is outdated and uses Gulp3 which limits it to run only on older node 10.x and also causes build issues in react app and CI build.

Merged in feature/PT-730-build (pull request #269)

Approved-by: Anna Caldwell