Add disease relevance filtering with user ID from token.
With the user id available from the token, the APIs that are dependent
on the user's configuration are now validated. This updates the disease
risk and events list to filter out any diseases that are not relevant.

The user location APIs no longer receive a user id in the URL, as this
is taken directly from the token. New constraint is added to limit user
locations to 50.

Updated Entity models to include the new Foreign Keys added to multiple
tables, as well as the EventDestinationAirport not including the
aggregate exportation risk (new table has this information).

Added province to geoname result for user location API.
Added reading of Environment variable to inject user ID for development.

Fixed path to zebra login.

Part of PT-730 and PT-734.
Merged in feature/PT-734-user-id-from-token (pull request #285)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>