Fix opening consecutive event details on event view
Event details on event view needs to be rendered in global view, i.e. no importation risk.

When in global view, geoname ID is not sent as part of the request. A bug in filtering events in EventController based on disease ID has also been fixed.

Part of PT-732
Merged in feature/PT-732-event-view (pull request #290)
Approved-by: Gursharn Singh