Event Query performance enhancements.
Previously, the Article and Location tables were left joined to events.
This was fine in the development environment due to there being a small
amount of each (m x n < 100). In production however, the amount of
articles can be as high as ~1800 and locations to be ~50. This causes
the amount of data returned (m x n) to be way too large to be
performant.

The queries have been updated to not do left joins from the Event table.
Instead, if locations or articles are included, then a separate query to
the DB will be called to include the minimal set of properties needed to
be returned.

Note that for the articles, there is special logic needed to be applied
for the DisplayName and SeqId properties, causing a huge amount of rows
to be still loaded into memory before applying the logic. Further
optimizations should be done by applying that logic on save of that data
from Surveillance and cached so that queries will return a much smaller
subset of data.

Also updated the Disease Risk API to return case counts when no geoname
id is provided. This is the aggregated case counts of all events of that
type of disease, with nesting applied.

Part of PT-799.
Merged in bugfix/PT-799-event-article-slowdown (pull request #308)
Approved-by: Basam Ahmad <basam.ahmad@bluedot.global>