Fix issues with map/disease-selection mismatch
Issue with map flickering when switching between diseases has also been fixed. This was caused by multiple calls to `showOutbreakExtent`.

API calls to get geoname shapes have also been modified to be done in chunks of 10 geonames to not have the map left blank for a long time.

Part of PT-809
Merged in bugfix/PT-809-outbreak-extent-on-map (pull request #320)
Approved-by: Gursharn Singh