Fix events filtering not applying relevance settings.
Previously, the Get Events controller used to ignore the disease id
passed in the query and always returned the set in the user relevance
settings. This was improperly fixed by a simple checking whether the
disease id was passed in, causing the disease relevance settings to
never be applied when a disease id was passed in.

The proper fix is that if both are available then both need to be
applied, i.e. if disease id is sent, also apply relevance settings on
that disease id, and not mutually exclusive.

Fixes PT-824.
Merged in bugfix/PT-824-filter-less-than-threshold (pull request #333)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>