Remove dependency on geoname id for destination airports.
In the airports query, the geoname id was never used to filter out the
airports. Also, the requirements state that destination airports are
independent of geoname ids, hence this parameter has been removed from
the API all the way down to the Query Builder.

Re-added legacy aggregate destination airport check.
The data has not been guaranteed to not have -1 for the legacy aggregate
so the check to exclude that row will be re-added.

Part of PT-868.
Merged in bugfix/PT-868-destination-airports-risk (pull request #335)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>