Fix issue with event details not being displayed when a filter is applied
Previously, disease and event lists filter out items that do not match the search text. This prevents the click trigger for that list item to not be initiated. This has been resolved by instead adding an attribute that defines its visibility.

Part of PT-872
Merged in bugfix/PT-872-clicking-filtered-list-item (pull request #349)
Approved-by: Gursharn Singh