Fix login/logout discrepancy between zebra and app.
On Zebra side, the application cookie has been properly set to the
expiration set in the Web.config file (7 days) instead of the default 14
days. This now matches the same duration as the _rid refresh token.
Also, when "Remember me" is checked, then the token expiry will properly
match that to the ASP.NET application cookie (if unchecked, it will be
Session expiry).

On Insights App side, the isLoggedIn function has been updated to use
the Refresh API to check whether the _rid exists, if not then it is
assumed that the user is no longer logged in. The token refresh code
also stores the cookie in there now, so that any refresh call will
automatically save the cookie, and not depend on the callback to save.

Code in the startup has been updated accordingly to handle the
isLoggedIn promise.

Fixes PT-827.
Merged in bugfix/PT-827-cookie-check (pull request #375)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>