Fix issue with child article toggle not working
By the time the script with the callback binding is executed, the element has not yet been loaded. This results in the event callback not being called. This is resolved by binding to the body instead and using a child selector for the actual event handling.

Fixes PT-938
Merged in bugfix/PT-938-surveillance-cluster-toggle (pull request #376)
Approved-by: Gursharn Singh