Mark not-calculated for events with only countries.
The model is never run on event locations of type Country. That means if
an event only has Countries for locations, then the model is not
actually run. Previously, the IsModelNotRun flag is false for these,
causing the UI to say Unlikely/Negligible, when the values are just the
default. Updated to update the flag if an event only has Country for
event locations.

Also refactored the Destination Airport Query builder to use the new
Foreign Keys and to also select the fields that are needed.

Fixes PT-947.
Merged in bugfix/PT-947-country-only-not-calculated (pull request #381)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>