Fix return url with single slash not redirecting.
When the return url is just a single slash, the redirect just returns to
the same page, which is the login page again. Added a check to not
redirect when it's just a single slash for the return url, and to
redirect to the app root page.

Also fixed the public layout not being used when the login was not
successful.

Fixes PT-951.
Merged in bugfix/PT-951-return-url-bug (pull request #387)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>