Handle timezone when calculating duration
Resolution to the level of minutes and hours is not necessary and can be approximated to today, as discussed with @James Yoon.

Calculate duration based on date part only.

UTC zone is being assumed for field Event{LastUpdatedDate} returned by api. Confirmed with @Basam Ahmad

Merged in bugfix/PT-956-insights---locstrat---event-updat (pull request #395)

Approved-by: Basam Ahmad