Modify Geoname Query to get the shape as text.
The Shape when there is a polygon within the multipolygon that is not
CCW, then NetTopologySuite considers it as a hole causing an exception.
To address this issue, we will load the shape with .AsText() to directly
convert to string from the SQL result, so that no conversion to Geometry
object is required.

NOTE: This does not fix the Fiji display issue on the map; it does
      prevent unexpected exception from occuring when the orientation of
      the shape is reversed.

Performance fix to not get local events when no geoname.
When no geoname is provided (global view or traditional view), the local
cases are not relevant, so added code to prevent calling SP when no
geoname is provided.

Fixes PT-869.
Merged in bugfix/PT-869-shape-query-as-text (pull request #405)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>