Fix bug where modelNotRun logic not applied everywhere.
Moved the IsModelNotRun logic into the class directly so that it can be
computed once and reused everywhere that needs this check.

The airports do not need to be loaded when no model is run. Also, the
aggregated risk depends on the fact that at least 1 event had their
model run, as well as the aggregation to exclude any events that are not
supposed to be included.

Fixes PT-972.
Merged in bugfix/PT-972-model-not-run (pull request #412)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>