Add column to store the WKT to CountryProvinceShapes table.
The shape AsText is used for shoging the map on the UI which is used a
lot more than the raw shape. Added this column to have quick retrieval
of the shape as the WKT instead of transforming at runtime, which is
slow for certain complex shapes.

Part of PT-997.
Merged in bugfix/PT-997-country-wkt-cache (pull request #439)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>