Break rendering of outbreak extent into chunks of 10 locations
This will allow the map not to appear frozen for such a long time when a disease has a lot of locations associated with it. Cancellation on event API has also been added to prevent previous calls to render on map.

Part of PT-1016
Merged in bugfix/PT-1016-map-rendering-in-chunks (pull request #438)
Approved-by: Gursharn Singh