Update entity framework to use table inclusion.
Updated the scaffold command to only include tables that are used, to
reduce the unused tables in the dbContext. Also updated
CountryProvinceShapes model to include the newly added shape as text
column.

Update the GeonameQueryBuilder to reduce column selection as well as
updated to take the new column for the shape as text field.

Fixes PT-997.
Merged in bugfix/PT-997-shape-text-column (pull request #441)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>