Implement Acquisition mode to disease api.
Also refactored to move Outbreak Potential Category business logic into
code so that the join is no longer necessary for the table.

Rewrote the Disease Query to select a smaller subset of fields, due to
the amount of joining that is now occuring with all the tables.

Part of PT-572
Merged in feature/PT-572-disease-details-api (pull request #476)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>