Implement service to communicate with Email Rendering.
Added models and services for Proximal, Weekly, and New Event, as well
as updated DI to allow the controllers to call these services. Updated
appsettings to include section for configuring the HttpClient that will
be used to make requests to the Email Rendering APIs.

Fixed the Email Rendering JS project to use camelCase instead of
PascalCase, which is what the objects are when serialized. Fixed the
data for query being a string while POST body being an object causing
JSON.parse to fail.

Removed Analytics configuration (again!) because analytics is configured
in the Email Rendering project, not injected via the Engine.

Part of PT-1078.
Merged in feature/PT-1078-render-api-call (pull request #506)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>