Implement logic for New Event email data.
Updated the function to include the geonameId that has been found to
intersect the location. Removed clean-up calls in the function as we
want all locations that have intersected, not just the first one.

Refactored and added NotificationService base class for common logic in
sending emails.

Implements PT-1132 and PT-1078.
Merged in feature/PT-1078-new-event-data-logic (pull request #515)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>