Fix unit tests by specifying DateTime.Kind for dates.
Unit tests have been failing randomly during the pipelines more
recently. Investigation discovered that the failing tests are showing
expected and actual of DateTime fields not matching, with the mismatch
happening all on 2020-03-08 between 2am to 3am. The mismatch is also
exactly 1 hour off (minutes and seconds are correct, only hour is
wrong).

The actual cause is due to the Daylight Savings happening during that
hour and there was some implicit conversion happening on the DateTime
object during the test where the initial creation is as is, but after
processing and stored via the ZebraUpdateController, the offset appears.

It seems that Json Serializer uses the DateTime.Kind to serialize the
DateTime object, where if it is set to Local it will add the appropriate
time zone offset, or if it is set to UTC it will add the Z time zone,
and if it is set to Unspecified it will leave it blank. Since the time
set between 2am and 3am of the daylight savings hour is not valid, the
hour is automatically adjusted when de/serialized.

Modified all DateTime objects used in the Unit Tests to explicitly set
the DateTime.Kind to Unspecified, to prevent unexpected time zone
adjustments during de/serialization.

Merged in bugfix/unit-test-random-fail (pull request #527)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>