Add testing mode flag for sending emails.
When in UAT testing, the flag to not actually send to the users is
enabled, which will direct all emails to be sent to the QA inbox
instead. Emails will also mention who the original recipient for that
email was for.

Updated MJML templates to include the conditional section, as well as
implemented helper function to allow rendering sub-mjml component with
handlebars. Updated notification settings to include flag and recipient
list for testing.

Part of PT-1078.
Merged in feature/PT-1078-testing-mode-flag (pull request #533)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>