Add PaidUsers role during user creation by default.
With the change where all users of Insights are paid by default, admins
creating user do not have to take an extra step to edit the user just to
add the PaidUsers role. Fixed by including the PaidUsers role as part of
the set of roles added to the user during creation.

Fixes PT-1026.
Merged in bugfix/PT-1026-default-paid-role (pull request #546)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>