Fix issue with outbreak extent not being displayed for some diseases
This issue is caused by the `useNonMobileEffect` implementation that creates a new array of dependencies at every call. Since a new dependency array is generated each time, React always sees them differently and calls the corresponding action. Because of this, all references to it are called regardless of whether there's a change in the dependencies or not. In the case of the outbreak extent not showing, `renderAois()` was being called every time a new disease is selected which cancels the previous `getGeonameShapes` request. Requests related to outbreak extent are chunked in groups 10 so if the disease has less than 10 geonames associated with it, that sole request will be cancelled and nothing would be rendered.

This commit also remove duplicate calls to `renderAois()` when user's locations are first loaded

Part of PT-1143
Merged in hotfix/PT-1143-outbreak-extent-not-displayed (pull request #544)
Approved-by: Kevin Chu