PT-1067 deep link routing for all sidebar panels
Now the following routing is supported:
/location
/location/24124
/location/24124/disease/24
/location/24124/disease/24/event/2415
/event
/event/2125

Most of the refactoring for this feature was to pass API-loaded data from individual panels up to the parent *View component that would designate how they are shared. The other nuance is the URL params that are passed as props to the *View component need to trigger an effect (see custom useDependentState hook) to covert to respective numeric IDs, otherwise a flicker happens on mobile while no panel is visible during a short period of time.

Merged in feature/PT-1067-deep-routing-for-insights-app (pull request #541)
Approved-by: Kevin Chu