Update config to try to read from environment variables.
The config.json cannot be injected in the deployment pipeline,
however we can inject environment variables to be in the scope of
the runtime environment. This is accessed via process.env object in
JavaScript.

Split the constant values from the configuration fields, and
updated so that config is read from environment before defaulting
to the JSON.

Also fixed old config key name in the test email.

Fixes PT-1160
https://docs.microsoft.com/en-us/azure/azure-functions/functions-reference-node#environment-variables
Merged in bugfix/PT-1160-email-renderer-config-variables (pull request #552)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>