Fix incorrect new event user filtering.
Previously, new event email only goes to user who are located within the
event locations. This implements only half the old email functionality
(there used to be 2 APIs, one for those proximal and those with >1%
risk). This caused a lot of issues, users that have risk don't get the
new event email, nor do all locations show up in the email when it's
unlikely.

Reworked the logic to have the starting point be all users, then filter
out if their relevant settings cause there to be no locations to be
shown.

Also renamed the Affected Users method to be more specific as that only
returns users that are geographically relevant, not risk-related.

Fixes PT-1176, PT-1178, and PT-1180.
Merged in bugfix/PT-1176-new-event-logic-fix (pull request #567)
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>