PT-939 - fix duplicates in location names in panels
First attempt was to add a "unique" filter to the atom array. However this bug is happening
because the name field for province and country its a redunant atom as it always point to the
top-level atom, already present in country and province fileds respectively.
I reworked it to return different atoms based on locationType.

* PT-939 - fix duplicate location names
* PT-939 - redo getLocationFullName based on locationType

Merged in bugfix/PT-939-minimized-location-duplicating (pull request #572)
Approved-by: Kevin Chu