Fix missing since date when no history.
When there is no history, then the updated location has the date set to
null, causing empty since date. The since date is now the earliest date
in the history table that has a difference to the current case counts.
If it doesn't exist, then default to the most recently sent proximal
email date. And if there never was a previous email sent, then default
to the event start date. Note that empty history should no longer
happen with the following change below.

Updated New Event service to update the history when sending email, as
this is when the first update is sent to users. Proximal emails are
subsequent updates after the first event email. Both of these now call
the SP that updates the event location to the history table with Type 1.

Added a check for New Event email to prevent sending it again if it has
been sent in the past.

Also updated MJML template to handle the case if the sent date is still
empty to not show "since".

Fixes PT-1197.
Merged in bugfix/PT-1197-missing-since-date (pull request #588)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>
https://bluedotglobal.atlassian.net/browse/PT-1066?focusedCommentId=12004&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-12004