Fix issue with non-proximal AOIs being listed in proximal e-mail as triggers
The user AOI list was being filtered based on all event locations not just the ones that triggered the update. This means that if an event has Canada and Toronto as event locations, all AOIs proximal to Canada were being included even if the update was just to Toronto.

Part of PT-1198
Merged in bugfix/PT-1198-proximal-notification-locations (pull request #589)
Approved-by: Kevin Chu