Remove filter flags when calling helper function.
The 3 flags sent to the helper function applies filtering on:
1. Must be a paid user
2. Must have New Event (Proximal) Notification Enabled
3. Must have email confirmed.

Previously, these flags are all set to true, so the helper function is
not generic. Fixed by sending all flags as false.

Moved the checks to the relevant places. The email confirmed check has
been applied as a base for all emails (this was a bug where this flag
was not checked). Refactored the 3 notification service to use the base
method to apply the necessary filters for users to receive emails.

Fixes related to PT-1198.
Merged in hotfix/PT-1198-fn-flags (pull request #597)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>
Approved-by: Sonya Karamchandani <sonya.karamchandani@bluedot.global>