Refactor Disease Service for injecting configurations.
This is part of an ongoing refactor to rework all the services to
make them more re-usable. The current problem is that the query
builders are configurable for which properties to load, but the
service themselves have fixed behaviour that always loads a lot of
properties that are not needed by the caller, causing slowdown in
performance if the service is re-used. This is especially true for
the Event Service which loads case counts.

To address this, we are giving the control of which properties to
load to the caller of the service. This is implemented with a builder
pattern which extracts a lot of the similar functionality out of the
Query Builder classes.

This change will first address the Disease related services. Updated
the callers to now pass in the built config. Already there are places
that are loading less information (e.g. Outbreak Potential only cares
about the OutbreakPotentialAttributeId).

Also added a String Extension method to default empty strings to null
or a parameter if provided. This cleans up the ternary operators in
the Disease Service when constructing display values for the disease
properties.

Refactor for PT-1164.
Merged in feature/PT-1164-refactor-disease-service (pull request #614)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>