Refactor Event and Airport services.
Merged the Source and Destination Airport configs into a single class
as the airport properties should be independent on which table it
came from. Reworked the SourceAirportJoinResult to selectively load
properties that are used, not include the whole EF object.

Also updated Airport Query Builders to use the Foreign Keys that were
added previously to not have to explicitly join station and geonames.

Refactored the Event Service and Event Query Builders with the new
Config Builder pattern. Moved loading logic for fields like outbreak
potential, or history into a configurable property in the config so
that the service will conditionally load it (as opposed to always
loading when called). This allows the event service for re-use as
the caller can now choose not to load a lot of the heavy operations.

Part of PT-1164.
Merged in feature/PT-1164-refactor-event-service (pull request #621)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>