Fix weekly case count not applying nesting.
The weekly event location case count was a direct sum over all event
locations, causing double counting of cases. This caused an overly
large value in the increased case count sent in the weekly emails.

Updated the WeeklyNotificationService to retrieve the event locations
for both current and history, apply nesting, then get the difference
before summing as the increased case count.

Also fixed a bug that was miscalculating the difference between the
two flattened case counts. The new method to calculate is to create
a delta dictionary for each location. For weekly email, we construct
the hierarchical structure back so that we only aggregate the top
level cases.

Fixes PT-1252.
Merged in hotfix/PT-1252-weekly-case-count-bug (pull request #626)
Approved-by: Anna Caldwell <anna.caldwell@bluedot.global>
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>