Move Update Weekly Data call into Weekly Email API.
Previously, the update weekly data call is separate from the weekly
email call. This was triggered in the WeeklyOperationConsole project
where once the request is completed for the email API call, it then
triggers the update API call. The problem was that any accidental
completion (whether network error or timeouts) will cause the code
to continue even when not all the emails were sent, and as a result
history data is pre-maturely updated.

Fixed by moving the SP call directly into the API (similar to what
was done in Proximal emails), so that the Email has control of when
to update the data, i.e. when all emails are sent.

Fixes PT-1265.
Merged in bugfix/PT-1265-weekly-email-data-update-being-t (pull request #655)
Approved-by: Predrag Ostojic <predrag.ostojic@bluedot.global>