Added unit tests to all helper classes.
Current coverage of the helpers in the Biod.Insights.Service project
is now at 100%, with it being only 9% of the whole project.

Added tests to ArticleHelper, and fixed logic to prevent exceptions
when the sourceUrl is null.

Added tests to OutbreakPotentialCategoryHelper.

Added tests to StringFormattingHelper and fixed logic in FormatPeriod
to prevent accessing a null min, max or average value. Also fixed
logic in FormatTime where checking for 1 is inaccurate when rounded,
(i.e. 1.04 minutes get rounded to 1 minute, but the formatting of the
plural form would be wrong). Also added tolerance for comparing with
floating point arithmetic.

Added tests to RiskCalculationHelper.

Added tests to ClaimsHelper. Refactored the ClaimsHelper to allow the
injection of the environment variables. There is now an interface and
class injected as a singleton for Insights API. For unit testing, a
separate implementation fo the interface is used.

Added tests to DiseaseRelevanceHelper. There was a bug where it was
not allowing risks equal to the threshold (only > 0.1 not >= 0.1).

Added tests to OrderingHelper.

Merged in unit-test-project (pull request #660)
Approved-by: Gursharn Singh <gursharn.singh@bluedot.global>